added animations for zaps
