+++ features/homefeed/components/UserBanner.js
@@ -2,43 +2,12 @@ import { View, Text, Pressable } from "react-native";
 import React from "react";
 import { Image } from "expo-image";
 import globalStyles from "../../../styles/globalStyles";
-import colors from "../../../styles/colors";
 import { useNavigation } from "@react-navigation/native";
 import Ionicons from "@expo/vector-icons/Ionicons";
-import Animated, {
-    withTiming,
-    useAnimatedStyle,
-    interpolateColor,
-    useDerivedValue,
-} from "react-native-reanimated";
 
 const UserBanner = ({ user, event, width, isZapped }) => {
     const imageDimensions = (width / 100) * 12;
     const navigation = useNavigation();
-
-    const bgProgress = useDerivedValue(() => {
-        return withTiming(isZapped ? 1 : 0);
-    });
-
-    const textStyle = useAnimatedStyle(() => {
-        const color = interpolateColor(
-            bgProgress.value,
-            [0, 1],
-            ["#ffffff", colors.backgroundPrimary]
-        );
-
-        return { color };
-    });
-
-    const addressStyle = useAnimatedStyle(() => {
-        const color = interpolateColor(
-            bgProgress.value,
-            [0, 1],
-            [colors.primary500, colors.backgroundPrimary]
-        );
-
-        return { color };
-    });
 
     return (
         <Pressable
@@ -70,22 +39,22 @@ const UserBanner = ({ user, event, width, isZapped }) => {
                 recyclingKey={event.pubkey}
             />
             <View style={{ flex: 1, marginLeft: 12 }}>
-                <Animated.Text
-                    style={[globalStyles.textBodyBold, { textAlign: "left" }, textStyle]}
+                <Text
+                    style={[globalStyles.textBodyBold, { textAlign: "left" }]}
                 >
                     {user?.name || event.pubkey.slice(0, 16)}
-                </Animated.Text>
-                <Animated.Text
+                </Text>
+                <Text
                     style={[
                         globalStyles.textBodyS,
-                        { textAlign: "left" }, addressStyle,
+                        { textAlign: "left" },
                     ]}
                 >
                     {user?.nip05}
                     <Ionicons
                         name={user?.nip05 ? "checkbox" : "close-circle"}
                     />
-                </Animated.Text>
+                </Text>
             </View>
         </Pressable>
     );
